Create Cost of Shipping
A shipping company calculates the cost (in dollars) of shipping based on the weight of the package (in pounds).

Write a program that prompts the user to enter the weight of the package and display the shipping cost. If the weight is greater than 20, display a message: "The package cannot be shipped''.